commit for jar execution with days included in database

+++ DeleteFiles/src/main/java/app/filemanager/FileService.java
@@ -13,33 +13,48 @@ import java.util.List;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import app.repository.FileRepositoryJDBC;
+import app.beans.*;
+
 @Service
-public class DeleteFileService {
+public class FileService {
 	
-	private static Logger LOG = LoggerFactory.getLogger(DeleteFileService.class);
+	private static Logger LOG = LoggerFactory.getLogger(FileService.class);
 
 //	private static final String DELETE_RUTH_NAME = "\\\\sarroca\\comu-inf$\\Suport\\test\\FILES_TO_DELETE";
-	private static final String MOVE_RUTH_NAME = "\\\\sarroca\\comu-inf$\\Suport\\test\\FILES_MOVED";
+//	private static final String MOVE_RUTH_NAME = "\\\\sarroca\\comu-inf$\\Suport\\test\\FILES_MOVED";
+	
+	@Autowired
+	private FileRepositoryJDBC repoFileJDBC;
+	
+	public int getDaysByReference(String reference) {
+		return repoFileJDBC.getDaysByReference(reference);
+	}
 
-	public void moveFiles(String path, String days) throws IOException {
+	public void moveFiles(String reference) throws IOException {
+		FBW300PATH action = repoFileJDBC.getActionByReference(reference);
+		
 		List<File> filesToMove = new ArrayList<>();
-		filesToMove = getOnlyFilesToWork(path, days);
-		moveOldFiles(path, filesToMove);
+		filesToMove = getOnlyFilesToWork(reference, action.getLimitdays());
+		moveOldFiles(action.getRootpath(),action.getMovepath(), filesToMove);
 	}
 	
 	
-	public void deleteFiles(String path, String days) throws IOException {
+	public void deleteFiles(String reference) throws IOException {
+		FBW300PATH action = repoFileJDBC.getActionByReference(reference);
+				
 		List<File> filesToDelete = new ArrayList<>();
-		filesToDelete = getOnlyFilesToWork(path, days);
-		deleteOldFiles(filesToDelete);
+		filesToDelete = getOnlyFilesToWork(action.getRootpath(), action.getLimitdays());
+		deleteOldFiles(filesToDelete, action.getRootpath());
 	}
 
-	private List<File> getOnlyFilesToWork(String path, String days) {
+	private List<File> getOnlyFilesToWork(String path, int limitdays) {
 		List<File> filesInPath = getFilesToWork(path);;
 		
-		long limitDate = getLimitDate(days);
+		long limitDate = getLimitDate(limitdays);
 		
 		List<File> onlyFilesToDelete = new ArrayList<>();
 
@@ -58,7 +73,7 @@ public class DeleteFileService {
 	}
 
 
-	private long getLimitDate(String days) {
+	private long getLimitDate(int limitdays) {
 		//Obtenir data actual
         Date currentDate = new Date();
         
@@ -67,7 +82,7 @@ public class DeleteFileService {
         calendar.setTime(currentDate);
 
         //Número díes a restar
-        int diasARestar = Integer.parseInt(days);
+        int diasARestar = limitdays;
         
         //Restar els díes a la data
         calendar.add(Calendar.DAY_OF_MONTH, -diasARestar);
@@ -78,20 +93,22 @@ public class DeleteFileService {
 		return nuevaFecha.getTime();
 	}
 
-	private void deleteOldFiles(List<File> filesToDelete) {
-		if(filesToDelete != null) {
+	private void deleteOldFiles(List<File> filesToDelete, String rootpath) {
+		if(filesToDelete != null && !filesToDelete.isEmpty()) {
 			for(File file: filesToDelete) {
 				if (file.delete()) {
 					LOG.warn("el fitxer eliminat"+ file.getName());
 				}
 			}
+		} else {
+			LOG.warn(" Al path "+rootpath+" no hi ha fitxers que compleixin les característiques per ser eliminats");
 		}
 	}
 	
-	private void moveOldFiles(String path, List<File> filesToMove) throws IOException {
+	private void moveOldFiles(String rootpath, String movePath, List<File> filesToMove) throws IOException {
 		for (File file : filesToMove) {
-			Path temp = Files.move(Paths.get(path + "\\" + file.getName()),
-					Paths.get(MOVE_RUTH_NAME + "\\" + file.getName()));
+			Path temp = Files.move(Paths.get(rootpath + "\\" + file.getName()),
+					Paths.get(movePath+ "\\" + file.getName()));
 			if (temp != null) {
 				LOG.warn("File moved successfully");
 			} else {
